🛠️Refactor: 전반적인 코드 수정
서적 검색 API에 인증 추가, 일부 테스트 진행 시 데이터베이스 초기화 작업 제거, 메서드 명 수정 등을 진행했습니다.
