+++ src/main/java/com/bombombom/devs/book/controller/BookController.java
@@ -30,7 +30,7 @@ public class BookController {
     }
 
     @PostMapping
-    public ResponseEntity<BookListResponse> bookAdd(
+    public ResponseEntity<BookListResponse> addBook(
         @Valid @RequestBody BookAddRequest bookAddRequest) {
         NaverBookApiResult naverBookApiResult = bookService.findBookUsingOpenApi(
             bookAddRequest.toServiceDto());
