+++ src/main/java/com/bombombom/devs/book/controller/dto/BookInfo.java
@@ -1,0 +1,14 @@
+package com.bombombom.devs.book.controller.dto;
+
+import lombok.Builder;
+
+@Builder
+public record BookInfo(
+    String title,
+    String author,
+    String publisher,
+    Long isbn,
+    String tableOfContents
+) {
+
+}
