+++ src/main/java/com/bombombom/devs/book/enums/SearchOption.java
@@ -1,10 +1,7 @@
 package com.bombombom.devs.book.enums;
 
 public enum SearchOption {
-    TOTAL("total"),
-    TITLE("title"),
-    AUTHOR("author");
-
-    SearchOption(String option) {
-    }
+    TOTAL,
+    TITLE,
+    AUTHOR
 }
