+++ src/main/java/com/bombombom/devs/book/naverapi/enums/SortType.java
@@ -1,9 +1,6 @@
 package com.bombombom.devs.book.naverapi.enums;
 
 public enum SortType {
-    SIM("sim"),     // 정확도순으로 내림차순 정렬
-    DATE("date");   // 출간일순으로 내림차순 정렬
-
-    SortType(String sort) {
-    }
+    SIM,
+    DATE
 }
