+++ src/test/java/com/bombombom/devs/book/service/BookServiceTest.java
@@ -27,9 +27,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.test.context.ActiveProfiles;
 
-@ActiveProfiles("test")
 @ExtendWith(MockitoExtension.class)
 public class BookServiceTest {
