+++ src/test/java/com/bombombom/devs/user/service/UserServiceTest.java
@@ -9,8 +9,10 @@ import com.bombombom.devs.user.service.dto.UserProfileResult;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -20,8 +22,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.when;
 
-@ActiveProfiles("test")
-@SpringBootTest
+@ExtendWith(MockitoExtension.class)
 public class UserServiceTest {
 
     private final String USERNAME = "bombombom";
@@ -40,9 +41,9 @@ public class UserServiceTest {
         Given
          */
         SignupCommand signupCommand = SignupCommand.builder()
-                .username(USERNAME)
-                .password(PASSWORD)
-                .build();
+            .username(USERNAME)
+            .password(PASSWORD)
+            .build();
 
         doReturn(true).when(userRepository).existsByUsername(any(String.class));
